requirements.txt
# Core Framework
flask==2.3.0
werkzeug==2.3.0

# Deep Learning
tensorflow==2.13.0
torch==2.0.1
scikit-learn==1.3.0

# Audio Processing
librosa==0.10.0
pydub==0.25.1
soundfile==0.12.1

# NLP & Text Processing
transformers==4.30.0
nltk==3.8.1

# Computer Vision
opencv-python>=4.8.0,<5.0.0
Pillow>=10.0.0,<11.0.0
face-recognition>=1.3.0

# Data Processing
numpy==1.24.0
pandas==2.0.0
pywavelets==1.4.1
joblib==1.3.0

# Database
sqlalchemy==2.0.0
pymysql==1.1.0

# Security
bcrypt==4.0.1
Flask-WTF==1.2.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0

# Visualization
matplotlib==3.7.0
seaborn==0.12.0

# Configuration
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# Testing
pytest==7.4.0
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development Tools (optional)
black==23.7.0
flake8==6.1.0
mypy==1.5.0
